requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
black==20.8b1
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2
dateparser==0.7.6
Django==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.11.2
djoser==2.0.5
fernet==1.0.1
humanize==3.1.0
idna==2.10
kombu==5.0.2
maya==0.6.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pendulum==2.1.2
prompt-toolkit==3.0.8
pyaes==1.6.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
regex==2020.10.15
requests==2.24.0
six==1.15.0
snaptime==0.2.4
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
